Use 'subprocess' module for all sub-processes.
The 'subprocess' is the modern and recommended way to interact with
sub-processes in Python: use it instead of `os.popen` for capturing the
initial environment, like it's already being done for running the bash
commands and capturing the modified environment.

Invoke `subprocess.check_output` with `universal_newlines` enable which
handles the decoding to string of the byte encoded output.

Introduce a constant to abstract the 'bash' executable.