Add `all` target and Fisherman installation note.
In the current implementation, running `make` will cause the
undesired effect of running `make install` by default causing
Fisherman to deploy bass to ~/.config/fish/functions when it
should not.

Even if Fisherman was able to patch this for the specific case
of bass, it seems like a good practice to separate the build and
install steps.