Fully support interactive commands that use stdout
The environment and aliases are outputted from bash on a pipe, and the
script is outputted on FD 3. This leaves bash with full control over
stdio. Bass will now only write to stdio if there's a Python exception.

This also simplifies the code. Since there's no need to handle stdout
specially, the environment is just the first line of the output from
bash, which means there's no need for a divider anymore.

The fish code has been refactored to use local variables and propagate
exit statuses more sanely.